Create separate package for schema handling
