+++ gson/src/main/java/com/google/gson/schema/JsonSchemaValidator.java
@@ -1,4 +1,9 @@
-package com.google.gson;
+package com.google.gson.schema;
+
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonPrimitive;
 
 import java.net.URI;
 import java.util.HashMap;
