+++ gson/src/main/java/module-info.java
@@ -7,8 +7,9 @@ module com.google.gson {
 	exports com.google.gson.annotations;
 	exports com.google.gson.reflect;
 	exports com.google.gson.stream;
+    exports com.google.gson.schema;
 
-	// Optional dependency on java.sql
+    // Optional dependency on java.sql
 	requires static java.sql;
 
 	// Optional dependency on jdk.unsupported for JDK's sun.misc.Unsafe
